feat: socket重连次数限制
