+++ index.html
@@ -10,8 +10,8 @@
     <link rel="shortcut icon" type="image/ico" href="favicon.ico" />
     <meta name="HandheldFriendly" content="true" />
     <script charset="utf-8" src="https://map.qq.com/api/js?v=2.exp&key=2LWBZ-FEQK6-KKYS2-M6WR4-PFGS5-RZBP3"></script>
-  <link href="./dist/index.6f39fb31.css" rel="stylesheet"></head>
+  <link href="/dist/index.88720f2a.css" rel="stylesheet"></head>
   <body>
     <div id="root"></div>
-  <script type="text/javascript" src="./dist/index.b6517c4e.js"></script></body>
+  <script type="text/javascript" src="/dist/index.aa717251.js"></script></body>
 </html>
