+++ src/mixins/bot.js
